massively simplified dataflow, delete button operational
